requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
langchain-community
langchain-ollama
langchain_huggingface
tqdm
sentence-transformers
faiss-cpu
chromadb
langchain-chorma
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn  
langserve
sse_starlette
arxiv
pymupdf
wikipedia